testing

+++ 0x05-linked_list_palindrome/0-is_palindrome.c
@@ -9,7 +9,7 @@
  * Return: the number of elements in a linked listint_t list.
  */
 
-size_t listint_len(const listint_t *h)
+int listint_len(const listint_t *h)
 {
 	int nodes = 0;
